feat: extract shared ProjectForm component and expand admin create project page
