+++ web/app/suggest/page.tsx
@@ -1,133 +1,20 @@
 'use client'
 
-import React, { useEffect, useState } from 'react'
+import { useEffect } from 'react'
 import { useRouter } from 'next/navigation'
 import Header from '@/components/Header'
-import Button from '@/components/Button'
-import Checkbox from '@/components/Checkbox'
-import Radio from '@/components/Radio'
-import SkillPicker from '@/components/SkillPicker'
-import FilterDropdown from '@/components/FilterDropdown'
-import { COUNTRY_OPTIONS, LOCAL_GROUPS } from '@/lib/filter-options'
+import ProjectForm from '@/components/ProjectForm'
 import { useAuth } from '@/lib/auth-context'
 import { useToast } from '@/lib/toast'
-import { apiRequest, ApiError } from '@/lib/api'
-
-interface SelectedSkill {
-  skillId: number
-  proficiencyLevel: string
-}
-
-interface Task {
-  title: string
-  description: string
-}
-
-const URGENCY_OPTIONS = [
-  { value: 'low',    label: 'Low - Nice to have' },
-  { value: 'medium', label: 'Medium - Should do soon' },
-  { value: 'high',   label: 'High - Urgent / time-sensitive' },
-]
-
-const PROJECT_TYPES = [
-  { value: '', label: 'Select a project type…' },
-  { value: 'sprint', label: 'Sprint (1-2 weeks) - Focused burst of work with clear deliverable' },
-  { value: 'container', label: 'Time-boxed (1-3 months) - Defined scope with end date' },
-  { value: 'ongoing', label: 'Ongoing - Continuous work without fixed end date' },
-  { value: 'one_off', label: 'One-off task - Single deliverable, minimal coordination' },
-]
-
 
 export default function SuggestPage() {
   const router = useRouter()
   const { user, loading } = useAuth()
   const toast = useToast()
 
-  const [title, setTitle] = useState('')
-  const [description, setDescription] = useState('')
-  const [projectType, setProjectType] = useState('')
-  const [hoursPerWeek, setHoursPerWeek] = useState('')
-  const [urgency, setUrgency] = useState('medium')
-  const [country, setCountry] = useState('')
-  const [localGroup, setLocalGroup] = useState('')
-  const [duration, setDuration] = useState('')
-  const [collaborationLink, setCollaborationLink] = useState('')
-  const [seekingHelp, setSeekingHelp] = useState(true)
-  const [wantToOwn, setWantToOwn] = useState(false)
-  const [skills, setSkills] = useState<SelectedSkill[]>([])
-  const [tasks, setTasks] = useState<Task[]>([{ title: '', description: '' }])
-  const [submitting, setSubmitting] = useState(false)
-  const [fieldErrors, setFieldErrors] = useState<Record<string, string>>({})
-
-  function clearFieldError(field: string) {
-    setFieldErrors(prev => {
-      if (!prev[field]) return prev
-      const { [field]: _, ...rest } = prev
-      return rest
-    })
-  }
-
-  function fe(field: string) {
-    return fieldErrors[field]
-  }
-
   useEffect(() => {
     if (!loading && !user) router.replace('/login')
   }, [user, loading, router])
-
-  function updateTask(index: number, field: keyof Task, value: string) {
-    setTasks(prev => prev.map((t, i) => i === index ? { ...t, [field]: value } : t))
-  }
-
-  function addTask() {
-    setTasks(prev => [...prev, { title: '', description: '' }])
-  }
-
-  function removeTask(index: number) {
-    if (tasks.length > 1) setTasks(prev => prev.filter((_, i) => i !== index))
-  }
-
-  async function handleSubmit(e: React.FormEvent<HTMLFormElement>) {
-    e.preventDefault()
-    const validTasks = tasks.filter(t => t.title.trim())
-    if (!validTasks.length) {
-      toast('At least one task with a title is required.', 'error')
-      return
-    }
-    setSubmitting(true)
-    try {
-      await apiRequest('/api/projects', {
-        method: 'POST',
-        body: JSON.stringify({
-          title: title.trim(),
-          description: description.trim(),
-          project_type: projectType || null,
-          time_commitment_hours_per_week: hoursPerWeek ? Number(hoursPerWeek) : null,
-          urgency,
-          country: country || null,
-          local_group: localGroup || null,
-          estimated_duration: duration.trim() || null,
-          collaboration_link: collaborationLink.trim() || null,
-          is_seeking_help: seekingHelp,
-          want_to_own: wantToOwn,
-          is_seeking_owner: !wantToOwn,
-          skill_ids: skills.map(s => s.skillId),
-          skill_required_map: Object.fromEntries(skills.map(s => [s.skillId, true])),
-          tasks: validTasks.map(t => ({ title: t.title.trim(), description: t.description.trim() || undefined })),
-        }),
-      })
-      toast('Project submitted for review! We\'ll be in touch.', 'success')
-      setTimeout(() => router.push('/dashboard?tab=proposed'), 2000)
-    } catch (err: unknown) {
-      if (err instanceof ApiError && err.fieldErrors) {
-        setFieldErrors(err.fieldErrors)
-        toast('Please fix the errors below.', 'error')
-      } else {
-        toast(err instanceof Error ? err.message : 'Failed to submit proposal', 'error')
-      }
-      setSubmitting(false)
-    }
-  }
 
   if (loading || !user) return null
 
@@ -136,259 +23,17 @@ export default function SuggestPage() {
       <Header />
       <main className="max-w-4xl mx-auto px-6 py-5 pb-15">
         <h1 role="heading">Suggest a Project</h1>
-        <p>Have an idea for something PauseAI UK should do? Propose it here! Our team will review it and, if approved, it'll be visible to all volunteers.</p>
-
-        <form className="bg-surface rounded-xl shadow p-6 mb-4 overflow-hidden wrap-break-word" onSubmit={handleSubmit}>
-          <div className="mb-5">
-            <label htmlFor="project-title" className="required">Project Title</label>
-            <input
-              id="project-title"
-              type="text"
-              value={title}
-              onChange={e => { setTitle(e.target.value); clearFieldError('title') }}
-              required
-              placeholder="A clear, descriptive name for the project"
-              aria-invalid={!!fe('title') || undefined}
-            />
-            {fe('title') && <p className="text-sm mt-1" style={{ color: 'var(--error)' }}>{fe('title')}</p>}
-          </div>
-
-          <div className="mb-5">
-            <label htmlFor="project-description" className="required">Description</label>
-            <aside className="bg-brand-bg border border-brand-border rounded-lg px-4 py-3 mb-2 text-sm">
-              <dl className="m-0 columns-2 max-sm:columns-1 gap-x-6">
-                <dt className="font-semibold text-brand-text break-after-avoid mt-2 first:mt-0">Goal &amp; Impact</dt>
-                <dd className="m-0 mb-1.5 text-text-light leading-snug break-inside-avoid">What problem does this solve? What does success look like and how would you measure it?</dd>
-                <dt className="font-semibold text-brand-text break-after-avoid mt-2">Approach &amp; Activities</dt>
-                <dd className="m-0 mb-1.5 text-text-light leading-snug break-inside-avoid">What are the key steps? Do you have a rough sense of how this would work?</dd>
-                <dt className="font-semibold text-brand-text break-after-avoid mt-2">Feasibility</dt>
-                <dd className="m-0 mb-1.5 text-text-light leading-snug break-inside-avoid">Main uncertainties or risks? Resources or dependencies needed? Existing work to build on?</dd>
-                <dt className="font-semibold text-brand-text break-after-avoid mt-2">Team &amp; Collaboration</dt>
-                <dd className="m-0 mb-1.5 text-text-light leading-snug break-inside-avoid">What skills and availability are ideal? How do you imagine the team working together?</dd>
-                <dt className="font-semibold text-brand-text break-after-avoid mt-2">Challenge Your Idea</dt>
-                <dd className="m-0 text-text-light leading-snug break-inside-avoid">What&apos;s the best argument against doing this? What&apos;s your Theory of Change?</dd>
-              </dl>
-            </aside>
-            <textarea
-              id="project-description"
-              rows={6}
-              value={description}
-              onChange={e => { setDescription(e.target.value); clearFieldError('description') }}
-              placeholder="Describe your project idea…"
-              required
-              minLength={10}
-              aria-invalid={!!fe('description') || undefined}
-            />
-            {fe('description')
-              ? <p className="text-sm mt-1" style={{ color: 'var(--error)' }}>{fe('description')}</p>
-              : <p className="text-sm text-text-light mt-1">The more detail you provide, the easier it is to find the right contributors and get started.</p>
-            }
-          </div>
-
-          <div className="mb-5">
-            <FilterDropdown
-              id="project-type"
-              label="Project Type"
-              ariaLabel="Select project type"
-              value={projectType}
-              options={PROJECT_TYPES}
-              onChange={v => { setProjectType(v); clearFieldError('project_type') }}
-            />
-            {fe('project_type')
-              ? <p className="text-sm mt-1" style={{ color: 'var(--error)' }}>{fe('project_type')}</p>
-              : <p className="text-sm text-text-light mt-1">This helps contributors understand the commitment involved</p>
-            }
-          </div>
-
-          <div className="grid grid-cols-2 gap-5 mb-5">
-            <div>
-              <label htmlFor="hours-per-week">Hours per Week</label>
-              <input
-                id="hours-per-week"
-                type="number"
-                min={1}
-                max={40}
-                placeholder="e.g., 5"
-                value={hoursPerWeek}
-                onChange={e => { setHoursPerWeek(e.target.value); clearFieldError('time_commitment_hours_per_week') }}
-                aria-invalid={!!fe('time_commitment_hours_per_week') || undefined}
-              />
-              {fe('time_commitment_hours_per_week')
-                ? <p className="text-sm mt-1" style={{ color: 'var(--error)' }}>{fe('time_commitment_hours_per_week')}</p>
-                : <p className="text-sm text-text-light mt-1">Estimated weekly time from each contributor</p>
-              }
-            </div>
-
-            <div>
-              <FilterDropdown
-                id="urgency"
-                label="Urgency"
-                ariaLabel="Select urgency"
-                value={urgency}
-                options={URGENCY_OPTIONS}
-                onChange={v => { setUrgency(v); clearFieldError('urgency') }}
-              />
-              {fe('urgency') && <p className="text-sm mt-1" style={{ color: 'var(--error)' }}>{fe('urgency')}</p>}
-            </div>
-          </div>
-
-          {['sprint', 'container'].includes(projectType) && (
-            <div className="mb-5">
-              <label htmlFor="duration">Estimated Duration</label>
-              <input
-                id="duration"
-                type="text"
-                placeholder="e.g., 6 weeks, 2 months"
-                value={duration}
-                onChange={e => { setDuration(e.target.value); clearFieldError('estimated_duration') }}
-                aria-invalid={!!fe('estimated_duration') || undefined}
-              />
-              {fe('estimated_duration')
-                ? <p className="text-sm mt-1" style={{ color: 'var(--error)' }}>{fe('estimated_duration')}</p>
-                : <p className="text-sm text-text-light mt-1">Roughly how long do you expect this to take?</p>
-              }
-            </div>
-          )}
-
-          <div className="mb-5">
-            <FilterDropdown
-              id="country"
-              label="Country / Region"
-              ariaLabel="Select country"
-              value={country}
-              options={COUNTRY_OPTIONS}
-              onChange={v => { setCountry(v); setLocalGroup(''); clearFieldError('country') }}
-              searchable
-            />
-            {fe('country')
-              ? <p className="text-sm mt-1" style={{ color: 'var(--error)' }}>{fe('country')}</p>
-              : <p className="text-sm text-text-light mt-1">Where is this project based? Select &quot;Remote&quot; if location-independent.</p>
-            }
-          </div>
-
-          {(LOCAL_GROUPS[country]?.length ?? 0) > 0 && (
-            <div className="mb-5">
-              <FilterDropdown
-                id="local-group"
-                label="Local Group"
-                ariaLabel="Select local group"
-                value={localGroup}
-                options={[
-                  { value: '', label: 'Not specific to a local group' },
-                  ...LOCAL_GROUPS[country].map(g => ({ value: g, label: `${country} - ${g}` })),
-                ]}
-                onChange={v => { setLocalGroup(v); clearFieldError('local_group') }}
-              />
-              {fe('local_group')
-                ? <p className="text-sm mt-1" style={{ color: 'var(--error)' }}>{fe('local_group')}</p>
-                : <p className="text-sm text-text-light mt-1">Is this project relevant to a particular local group?</p>
-              }
-            </div>
-          )}
-
-          <div className="mb-5">
-            <label htmlFor="collaboration-link">Collaboration Doc / Link (optional)</label>
-            <input
-              id="collaboration-link"
-              type="text"
-              placeholder="e.g., https://docs.google.com/… or 'Will create a shared doc once team forms'"
-              value={collaborationLink}
-              onChange={e => { setCollaborationLink(e.target.value); clearFieldError('collaboration_link') }}
-              aria-invalid={!!fe('collaboration_link') || undefined}
-            />
-            {fe('collaboration_link')
-              ? <p className="text-sm mt-1" style={{ color: 'var(--error)' }}>{fe('collaboration_link')}</p>
-              : <p className="text-sm text-text-light mt-1">A URL to a planning doc or workspace, or just describe your plans for collaboration</p>
-            }
-          </div>
-
-          <div className="mb-5">
-            <label>Skills Needed</label>
-            <p className="text-sm text-text-light mt-0 mb-2">What skills would be helpful for this project?</p>
-            <SkillPicker value={skills} onChange={setSkills} />
-          </div>
-
-          <div className="mb-5">
-            <p className="font-medium mb-2">This project needs:</p>
-            <div className="flex flex-col gap-2 mb-4">
-              <Checkbox
-                checked={seekingHelp}
-                onChange={e => setSeekingHelp(e.target.checked)}
-              >
-                Help / contributors
-              </Checkbox>
-            </div>
-            <p className="font-medium mb-2">Project ownership:</p>
-            <div className="flex flex-col gap-2">
-              <Radio
-                name="ownership"
-                checked={!wantToOwn}
-                onChange={() => setWantToOwn(false)}
-              >
-                This project needs an owner / lead
-              </Radio>
-              <Radio
-                name="ownership"
-                checked={wantToOwn}
-                onChange={() => setWantToOwn(true)}
-              >
-                <span><strong>I want to lead this project</strong> &mdash; I&apos;ll be the owner and coordinate the work</span>
-              </Radio>
-            </div>
-          </div>
-
-          <div className="mb-5">
-            <label>Initial Tasks</label>
-            <p className="text-sm text-text-light mt-0 mb-2">
-              Break the project into at least one concrete task. This helps reviewers understand the scope and gives volunteers something to pick up.
-            </p>
-            {tasks.map((task, i) => (
-              <div key={i} className="bg-brand-bg rounded-lg p-4 mb-3 border border-brand-border">
-                <div className="mb-3">
-                  <label htmlFor={`task-title-${i}`} className="text-sm required">Task title</label>
-                  <input
-                    id={`task-title-${i}`}
-                    type="text"
-                    aria-label="Task title"
-                    placeholder="e.g. Draft copy for homepage"
-                    value={task.title}
-                    onChange={e => updateTask(i, 'title', e.target.value)}
-                  />
-                </div>
-                <div className="mb-2">
-                  <label htmlFor={`task-desc-${i}`} className="text-sm">Details (optional)</label>
-                  <textarea
-                    id={`task-desc-${i}`}
-                    placeholder="More detail about what needs doing…"
-                    value={task.description}
-                    onChange={e => updateTask(i, 'description', e.target.value)}
-                    style={{ minHeight: 60 }}
-                  />
-                </div>
-                {tasks.length > 1 && (
-                  <Button
-                    type="button"
-                    variant="ghost"
-                    icon
-                    onClick={() => removeTask(i)}
-                    aria-label="Remove task"
-                  >×</Button>
-                )}
-              </div>
-            ))}
-            <Button type="button" variant="secondary" onClick={addTask}>
-              + Add another task
-            </Button>
-          </div>
-
-          <div className="flex items-center gap-3 p-4 rounded-lg mb-5 bg-[#DBEAFE] text-[#1E40AF] border border-[#93C5FD] dark:bg-[#1E3A5F] dark:text-[#93C5FD] dark:border-[#2563EB]">
-            Your project will be reviewed by PauseAI UK team leads before being published. We&apos;ll reach out if we have questions or suggestions.
-          </div>
-
-          <Button type="submit" className="w-full" disabled={submitting}>
-            {submitting ? 'Submitting…' : 'Submit Project Proposal'}
-          </Button>
-        </form>
+        <p>Have an idea for something PauseAI UK should do? Propose it here! Our team will review it and, if approved, it&apos;ll be visible to all volunteers.</p>
+        <ProjectForm
+          action="/api/projects"
+          submitLabel="Submit Project Proposal"
+          showReviewNotice
+          requireTasks
+          onSuccess={() => {
+            toast('Project submitted for review! We\'ll be in touch.', 'success')
+            setTimeout(() => router.push('/dashboard?tab=proposed'), 2000)
+          }}
+        />
       </main>
     </>
   )
